+++ src/interfaces.ts
@@ -7,6 +7,7 @@ export interface HapInstance {
   username: string;
   connectionFailedCount: number;
   services: ServiceType[];
+  configurationNumber: number;
 }
 
 export interface HapEvInstance {
@@ -68,7 +69,9 @@ export interface ServiceType {
   accessoryInformation: any;
   refreshCharacteristics?: () => Promise<ServiceType>;
   setCharacteristic?: (iid: number, value: number | string | boolean) => Promise<ServiceType>;
+  setCharacteristicByType?: (type: string, value: number | string | boolean) => Promise<ServiceType>;
   getCharacteristic?: (type: string) => CharacteristicType;
+  getResource?: (body: ResourceRequestType) => Promise<ServiceType>;
   values: any;
   instance: HapInstance;
   uniqueId?: string;
@@ -102,4 +105,11 @@ export interface AccessoryInformationProperties {
   'Name': string;
   'Serial Number': string;
   'Firmware Revision': string;
+}
+
+export interface ResourceRequestType {
+  aid?: number,
+  "resource-type": string,
+  "image-width": number,
+  "image-height": number
 }